+++ lib/eve_online.rb
@@ -72,6 +72,8 @@ require "eve_online/esi/corporation_loyalty_store_offers"
 require "eve_online/esi/character_orders"
 require "eve_online/esi/corporation_orders"
 require "eve_online/esi/market_history"
+require "eve_online/esi/market_groups"
+require "eve_online/esi/market_group"
 
 require "eve_online/esi/character_attributes"
 require "eve_online/esi/character_skill_queue"
@@ -152,6 +154,7 @@ require "eve_online/esi/models/home_location"
 require "eve_online/esi/models/server_status"
 
 require "eve_online/esi/models/market_history"
+require "eve_online/esi/models/market_group"
 
 require "eve_online/esi/models/corporation"
 require "eve_online/esi/models/corporation_alliance_history"